requirements.txt
pydantic~=1.10.6
sqlalchemy~=2.0.5.post1
pandas~=1.5.3
requests~=2.28.2

pytz~=2022.7.1
psycopg2-binary
load_dotenv
injectable